chore: changed the file structured in NeuronStatus component
